bricks: Try to sort test cases in shelltest more intelligently.
This includes the possibility of using --sort-hint regex to select part of test
name for sorting (can be given multiple times and hints are applied left to
right; it can contain single match group which will be used as the hint if
available). This also includes trying to sort tests whose names seems to be
extension of other test name after the extended test, i.e. test.c is sorted
before test-special.c.